Separate electron tray to tray.js
