+++ main.js
@@ -4,12 +4,10 @@ const fs       = require('fs');
 const electron = require('electron');
 const appMenu  = require('./menu');
 const config   = require('./config');
+const tray     = require('./tray');
 
 const app      = electron.app;
 const ipcMain  = electron.ipcMain;
-const shell    = electron.shell;
-const Tray     = electron.Tray;
-const Menu     = electron.Menu;
 
 require('electron-debug')();
 
@@ -17,7 +15,6 @@ const BrowserWindow = electron.BrowserWindow;
 
 let mainWindow;
 let isQuitting = false;
-let tray = null;
 
 const isAlreadyRunning = app.makeSingleInstance(() => {
   if (mainWindow) {
@@ -38,8 +35,6 @@ function createMainWindow() {
   const isDarkMode = config.get('darkMode');
   const userAgent = 'Mozilla/5.0 (iPhone; CPU iPhone OS 9_1 like Mac OS X) AppleWebKit/601.1.46 (KHTML, like Gecko) Version/9.0 Mobile/13B143 Safari/601.1';
   const rammeDesktopIcon = path.join(__dirname, 'static/icon.png');
-  const rammeTrayDefaultIcon = path.join(__dirname, 'static/icon-18x18.png');
-  const rammeTrayWindowsIcon = path.join(__dirname, 'static/icon.ico');
   const maxWidthValue = 550;
   const minWidthValue = 400;
 
@@ -84,42 +79,7 @@ function createMainWindow() {
 
   win.on('page-title-updated', e => {
     e.preventDefault();
-    let icon = rammeTrayDefaultIcon;
-
-    if (process.platform === 'win32') {
-      icon = rammeTrayWindowsIcon;
-    }
-
-    // Create toolbar
-    tray = new Tray(icon);
-
-    const trayMenu = [{
-      label: 'Show / Hide',
-      click() {
-        !win.isMinimized() ? win.minimize() : win.show();
-      }
-    }, {
-      label: 'GitHub',
-      click() {
-        shell.openExternal('https://github.com/terkelg/ramme');
-      }
-    }, {
-      label: 'Issue',
-      click() {
-        shell.openExternal('https://github.com/terkelg/ramme/issues');
-      }
-    }, {
-      label: 'Quit',
-      click() {
-        app.quit();
-      }
-    }];
-
-    tray.setToolTip('ramme');
-    tray.setContextMenu(Menu.buildFromTemplate(trayMenu));
-    tray.on('click', () => {
-      win.isVisible() ? win.hide() : win.show();
-    });
+    tray.create(win);
   });
 
   return win;
@@ -134,7 +94,6 @@ function sendAction(action) {
 
   win.webContents.send(action);
 }
-
 
 app.on('ready', () => {
   electron.Menu.setApplicationMenu(appMenu);
